requirements.txt
discord.py==2.5.2
alembic==1.16.4
asyncpraw==7.8.1
asyncprawcore==2.4.0
asyncpg==0.30.0
black==25.1.0
isort==6.0.1
autoflake==2.3.1
vulture==2.14
SQLAlchemy==2.0.41
psycopg2==2.9.10
python-dotenv==1.1.1
aiofiles==24.1.0